# Directly apply `ipiv` permutation from `LU` without `Base.permute!!`