--- a/src/linsolve.jl
+++ b/src/linsolve.jl
@@ -1,15 +1,15 @@
 const PFac = PositiveFactorizations
 
-function apply_ipiv_perm!(w::AbstractVector, v::AbstractVector{<:Integer})
-    @inbounds for i in eachindex(v)
-        w[i], w[v[i]] = w[v[i]], w[i]
+function apply_ipiv_perm!(w::AbstractVector, p::AbstractVector{<:Integer})
+    @inbounds for i in eachindex(p)
+        w[i], w[p[i]] = w[p[i]], w[i]
     end
     return w
 end
 
-function luupdate!(lu::LU, p::AbstractVector, w::AbstractVector, v::AbstractVector)
+function luupdate!(lu::LU, w::AbstractVector, v::AbstractVector)
     m, n = size(lu)
-    apply_ipiv_perm!(w, getfield(lu, :ipiv), m)
+    apply_ipiv_perm!(w, getfield(lu, :ipiv))
     F = getfield(lu, :factors)
     @inbounds for i in 1:min(m, n)
         wi = w[i]
@@ -37,7 +37,6 @@ end
 mutable struct DenseLUSolver{F}
     ws::LUWs
     fac::F
-    perm::Vector{Int}
 end
 
 default_linsolver(fdf::OnceDifferentiable{V, M, V}, x0::V,
@@ -51,7 +50,7 @@ function init(::Type{DenseLUSolver}, J::AbstractMatrix, f::AbstractVector)
     ws = LUWs(J)
     # Do not destroy the original J
     fac = LU(LAPACK.getrf!(ws, copy(J))...)
-    return DenseLUSolver(ws, fac, Vector{Int}(undef, size(J, 1)))
+    return DenseLUSolver(ws, fac)
 end
 
 init(s::DenseLUSolver, fdf::OnceDifferentiable, x0::AbstractVector) =
@@ -61,7 +60,7 @@ update!(s::DenseLUSolver{<:LU}, J::AbstractMatrix) =
     (s.fac = LU(LAPACK.getrf!(s.ws, copyto!(getfield(s.fac, :factors), J))...); nothing)
 
 update!(s::DenseLUSolver{<:LU}, J::AbstractMatrix, w::AbstractVector, v::AbstractVector) =
-    (BLAS.ger!(one(eltype(J)), w, v, J); luupdate!(s.fac, s.perm, w, v); nothing)
+    (BLAS.ger!(one(eltype(J)), w, v, J); luupdate!(s.fac, w, v); nothing)
 
 solve!(s::DenseLUSolver{<:LU}, Y, J, F) = ldiv!(Y, s.fac, F)